# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthMeResponse": {
        "description": "Current app authentication state.",
        "properties": {
          "auth_method": {
            "anyOf": [
              {
                "enum": [
                  "api_key",
                  "plex_session",
                  "dev_bypass"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Method"
          },
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "is_admin": {
            "default": false,
            "title": "Is Admin",
            "type": "boolean"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthUser"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "authenticated"
        ],
        "title": "AuthMeResponse",
        "type": "object"
      },
      "AuthUser": {
        "description": "Current signed-in Plex user.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_admin": {
            "default": false,
            "title": "Is Admin",
            "type": "boolean"
          },
          "plex_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plex Id"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plex_id",
          "username"
        ],
        "title": "AuthUser",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlexLoginCompleteRequest": {
        "description": "Complete a pending Plex PIN login challenge.",
        "properties": {
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "title": "PlexLoginCompleteRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Plex Manager",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke the current Plex browser session and clear auth cookies.",
        "operationId": "logout_endpoint_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Logout Endpoint",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/plex/complete": {
      "post": {
        "description": "Complete a Plex PIN challenge, verify owner access, and issue cookies.",
        "operationId": "plex_complete_endpoint_api_v1_auth_plex_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlexLoginCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Plex Complete Endpoint",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}